- refactor first download test
